+++ packages/file-types/zarr/src/anndata-loaders/FeatureLabelsAnndataLoader.js
@@ -40,12 +40,16 @@ export default class FeatureLabelsAnndataLoader extends AbstractTwoStepLoader {
   }
 
   async load() {
+    const { path } = this.options;
     const superResult = await super.load().catch(reason => Promise.resolve(reason));
     if (superResult instanceof AbstractLoaderError) {
       return Promise.reject(superResult);
     }
     return Promise.all([
-      this.dataSource.loadVarIndex(),
+      // Pass in the obsEmbedding path,
+      // to handle the MuData case where the obsIndex is located at
+      // `mod/rna/index` rather than `index`.
+      this.dataSource.loadVarIndex(path),
       this.loadLabels(),
     ]).then(([featureIndex, featureLabels]) => Promise.resolve(new LoaderResult(
       {