+++ packages/file-types/zarr/src/utils.js
@@ -1,0 +1,5 @@
+export function dirname(path) {
+  const arr = path.split('/');
+  arr.pop();
+  return arr.join('/');
+}